fix: model id type

+++ src/model/article_tag.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ArticleTagModel {
-    pub article_id: i64,
-    pub tag_id: i64,
+    pub article_id: i32,
+    pub tag_id: i32,
 }
